+++ www/go/core/customfield/DateTime.php
@@ -25,7 +25,7 @@ class DateTime extends Base {
    */
 	public function defineFilter(Filters $filters) {		
 		
-		$filters->addDate($this->field->databaseName, function(Criteria $criteria, $comparator, $value, Query $query, array $filter){
+		$filters->addDateTime($this->field->databaseName, function(Criteria $criteria, $comparator, $value, Query $query, array $filter){
 			$this->joinCustomFieldsTable($query);						
 			$criteria->where('customFields.' . $this->field->databaseName, $comparator, $value);
 		});